Change backup format, group by profile instead of date
Backups are now grouped by profile instead of date. This means less
noise when running daily backup of a specific profile, earlier it
would've polluted the backups directory, now it'll add the backups to
the profile specific folder.

Also, this includes more information & users can run leo multiple
times a day. Earlier running more than once a day would've
over-written the previous backup, from now on it'll create a new file
because we include time in the filename.

Time will be in GMT.